app/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
influxdb-client==1.38.0
pandas==2.1.0
numpy==1.26.0
scipy==1.11.3
requests==2.31.0
waitress==2.1.2
Pillow==10.1.0
qrcode[pil]==7.4.2
beautifulsoup4==4.12.2
pybeerxml==1.0
recipe-scrapers==14.53.0
PyGithub==2.1.1